@@ -8,10 +8,11 @@
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.6.1",
     "axios": "^1.8.3",
-    "react": "^18.3.1",
+    "react": "^19.0.0",
     "react-cookie": "^8.0.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.0.0",
     "react-redux": "^8.0.2",
+    "react-router": "^7.5.0",
     "react-router-dom": "^5.3.0",
     "web-vitals": "^2.1.4"
   },
